fix: share valuation runtime settings across services
